Move non-pages into components directory
